requirements.txt
django==5.0.2
mysqlclient==2.2.4
python-dotenv==1.0.1
requests==2.31.0
djangorestframework==3.14.0
django-cron==0.6.0